+++ templates/part.listDeviceGroups.php
@@ -4,8 +4,6 @@
 	<th></th>
 	<th id="id"><span class="th-header">#</span></th>
 	<th id="deviceGroupName"><span class="th-header">Device Group</span></th>
-	<th id="assigendDevices"><span class="th-header">Assigned devices</span></th>
-	<th><span class="th-header"></span></th>
 	</thead>
 	<tbody>
 <?php foreach($_['deviceGroups'] as $deviceGroup){ ?>
@@ -19,8 +17,6 @@
 		</td>
 		<td class="td-data"><?php p($deviceGroup['id']); ?></td>
 		<td class="td-data"><?php p($deviceGroup['device_group_name']); ?></td>
-		<td class="td-data"></td>
-		<td class="td-data">Show devices</td>
 	</tr>
 <?php } ?>
 	</tbody>